fix imports to execute xi18n command for angular10
