+++ libs/core/src/lib/action-sheet/public_api.ts
@@ -3,3 +3,4 @@ export * from './action-sheet.component';
 export * from './action-sheet-item/action-sheet-item.component';
 export * from './action-sheet-control/action-sheet-control.component';
 export * from './action-sheet-body/action-sheet-body.component';
+export * from './action-sheet-mobile/action-sheet-mobile.module';
