+++ libs/core/src/lib/grid-list/components/grid-list-title-bar-spacer/grid-list-title-bar-spacer.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { ToolbarSpacerComponent } from '../../../toolbar/public_api';
+import { ToolbarSpacerComponent } from '../../../toolbar/toolbar-spacer.component';
 
 @Component({
     selector: 'fd-grid-list-title-bar-spacer',
