+++ libs/core/src/lib/grid-list/components/grid-list/grid-list-selection-mode.ts
@@ -1,0 +1,7 @@
+export type GridListSelectionMode =
+    | 'none'
+    | 'delete'
+    | 'singleSelect'
+    | 'singleSelectLeft'
+    | 'singleSelectRight'
+    | 'multiSelect';
