+++ libs/core/src/lib/grid-list/components/grid-list/index.ts
@@ -1,1 +1,2 @@
 export * from './grid-list.component';
+export * from './grid-list-selection-mode';
