--- libs/core/src/lib/grid-list/components/index.ts
@@ -1,35 +1,0 @@
-import { GridListComponent } from './grid-list';
-import { GridListItemComponent } from './grid-list-item';
-import { GridListTitleBarAdditionalTitleItemDirective, GridListTitleBarComponent } from './grid-list-title-bar';
-import { GridListFilterBarComponent } from './grid-list-filter-bar';
-import { GridListMoreBtnComponent } from './grid-list-more-btn';
-import { GridListFooterComponent } from './grid-list-footer';
-import { GridListItemFooterBarComponent } from './grid-list-item-footer-bar';
-import { GridListItemToolbarComponent } from './grid-list-item-toolbar';
-import { GridListGroupHeaderComponent } from './grid-list-group-header';
-import { GridListTitleBarSpacerComponent } from './grid-list-title-bar-spacer';
-
-export * from './grid-list';
-export * from './grid-list-item';
-export * from './grid-list-title-bar';
-export * from './grid-list-filter-bar';
-export * from './grid-list-more-btn';
-export * from './grid-list-footer';
-export * from './grid-list-item-footer-bar';
-export * from './grid-list-item-toolbar';
-export * from './grid-list-group-header';
-export * from './grid-list-title-bar-spacer';
-
-export const COMPONENTS = [
-    GridListComponent,
-    GridListItemComponent,
-    GridListTitleBarComponent,
-    GridListFilterBarComponent,
-    GridListMoreBtnComponent,
-    GridListFooterComponent,
-    GridListItemFooterBarComponent,
-    GridListItemToolbarComponent,
-    GridListGroupHeaderComponent,
-    GridListTitleBarSpacerComponent,
-    GridListTitleBarAdditionalTitleItemDirective
-];
