--- libs/core/src/lib/grid-list/index.ts
@@ -1,1 +1,0 @@
-export * from './public_api';
