+++ libs/core/src/lib/grid-list/public_api.ts
@@ -1,2 +1,12 @@
+export * from './components/grid-list/grid-list.component';
+export * from './components/grid-list-filter-bar/grid-list-filter-bar.component';
+export * from './components/grid-list-footer/grid-list-footer.component';
+export * from './components/grid-list-group-header/grid-list-group-header.component';
+export * from './components/grid-list-item/grid-list-item.component';
+export * from './components/grid-list-item-footer-bar/grid-list-item-footer-bar.component';
+export * from './components/grid-list-item-toolbar/grid-list-item-toolbar.component';
+export * from './components/grid-list-more-btn/grid-list-more-btn.component';
+export * from './components/grid-list-title-bar/grid-list-title-bar.component';
+export * from './components/grid-list-title-bar/grid-list-title-bar.directive';
+export * from './components/grid-list-title-bar-spacer/grid-list-title-bar-spacer.component';
 export * from './grid-list.module';
-export * from './components';
