+++ libs/core/src/lib/product-switch/public_api.ts
@@ -2,6 +2,3 @@ export * from './product-switch.module';
 export * from './product-switch/product-switch.component';
 export * from './product-switch-body/product-switch.item';
 export * from './product-switch-body/product-switch-body.component';
-export * from '../utils/drag-and-drop/drag-and-drop.module';
-export * from '../utils/drag-and-drop/dnd-item/dnd-item.directive';
-export * from '../utils/drag-and-drop/dnd-list/dnd-list.directive';
