+++ libs/core/src/lib/wizard/constants.ts
@@ -1,0 +1,4 @@
+export const ACTIVE_STEP_STATUS = 'active';
+export const CURRENT_STEP_STATUS = 'current';
+export const UPCOMING_STEP_STATUS = 'upcoming';
+export const COMPLETED_STEP_STATUS = 'completed';
