+++ libs/platform/src/index.ts
@@ -13,7 +13,6 @@ export * from './lib/domain/public_api';
 export * from './lib/components/info-label/public_api';
 export * from './lib/components/panel/public_api';
 export * from './lib/components/object-status/public_api';
-export * from './lib/components/form/combobox';
 export * from './lib/components/list/public_api';
 export * from './lib/components/thumbnail/public_api';
 export * from './lib/components/object-marker/public_api';
@@ -22,7 +21,7 @@ export * from './lib/components/table/public_api';
 export * from './lib/components/dynamic-page/public_api';
 export * from './lib/components/value-help-dialog/public_api';
 export * from './lib/components/feed-input/public_api';
-export * from './lib/components/upload-collection';
-export * from './lib/utils/pipes/pipe.module';
+export * from './lib/components/upload-collection/public_api';
+export * from './lib/utils/pipes/public_api';
 export * from './lib/components/approval-flow/public_api';
-export * from './lib/components/slider';
+export * from './lib/components/slider/public_api';
