+++ libs/platform/src/lib/components/form/combobox/index.ts
@@ -1,6 +1,6 @@
-export * from './combobox.module';
-export * from './combobox/combobox.component';
 export * from './commons/base-combobox';
+export * from './combobox/combobox.component';
+export * from './combobox.module';
 export * from './combobox-mobile/combobox/combobox-mobile.component';
 export * from './combobox-mobile/combobox-mobile.module';
 export * from './pipes/highlight.pipe';
