+++ libs/platform/src/lib/components/form/combobox/public_api.ts
@@ -1,0 +1,6 @@
+export { ComboboxSelectionChangeEvent } from './commons/base-combobox';
+export { ComboboxComponent } from './combobox/combobox.component';
+export { PlatformComboboxModule } from './combobox.module';
+export { ComboboxMobileComponent } from './combobox-mobile/combobox/combobox-mobile.component';
+export { PlatformComboboxMobileModule } from './combobox-mobile/combobox-mobile.module';
+export { HighlightPipe } from './pipes/highlight.pipe';
