+++ libs/platform/src/lib/components/form/multi-input/public_api.ts
@@ -1,0 +1,5 @@
+export { MultiInputSelectionChangeEvent } from './base-multi-input';
+export { PlatformMultiInputComponent } from './multi-input.component';
+export { PlatformMultiInputModule } from './multi-input.module';
+export { PlatformMultiInputMobileComponent } from './multi-input-mobile/multi-input-mobile.component';
+export { PlatformMultiInputMobileModule } from './multi-input-mobile/multi-input-mobile.module';
