+++ libs/platform/src/lib/components/form/public_api.ts
@@ -2,7 +2,6 @@ export * from './checkbox/checkbox.module';
 export * from './checkbox/checkbox.component';
 export * from './checkbox-group/checkbox-group.module';
 export * from './checkbox-group/checkbox-group.component';
-export * from './combobox';
 export * from './form-control';
 export * from './base.input';
 export * from './form-options';
@@ -33,9 +32,9 @@ export * from './time-picker/time-picker.module';
 export * from './time-picker/time-picker.component';
 export * from './step-input/public_api';
 export * from './input-group/public_api';
-export * from './combobox';
-export * from './multi-input';
+export * from './combobox/public_api';
+export * from './multi-input/public_api';
 export * from './platform-file-uploader/platform-file-uploader.module';
 export * from './platform-file-uploader/platform-file-uploader.component';
 export * from './auto-complete/auto-complete.module';
-export * from './select';
+export * from './select/public_api';
