+++ libs/platform/src/lib/components/form/select/public_api.ts
@@ -1,0 +1,4 @@
+export { FdpSelectData, FdpSelectionChangeEvent } from './commons/base-select';
+export { SelectComponent } from './select/select.component';
+export { OptionComponent } from './option/option.component';
+export { PlatformSelectModule } from './select.module';
