+++ libs/platform/src/lib/components/menu/public_api.ts
@@ -1,1 +1,4 @@
 export * from './menu.module';
+export * from './menu.component';
+export * from './menu-item.component';
+export * from './menu-trigger.directive';
