--- libs/platform/src/lib/components/upload-collection/index.ts
@@ -1,1 +1,0 @@
-export * from './public_api';
