+++ libs/platform/src/lib/components/upload-collection/public_api.ts
@@ -1,8 +1,8 @@
-export * from './upload-collection.module';
 export * from './services';
 export * from './dialogs';
+export * from './domain';
 export * from './directives';
 export * from './upload-collection/upload-collection.component';
 export * from './models/upload-collection-events.models';
 export * from './models/upload-collection.models';
-export * from './domain';
+export * from './upload-collection.module';
