+++ libs/platform/src/lib/domain/index.ts
@@ -3,4 +3,3 @@ export * from './data-source';
 export * from './data-model';
 export * from './array-data-source';
 export * from './base-data-provider';
-export * from './public_api';
