+++ libs/platform/src/lib/utils/pipes/pipe.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 
-import { ConvertBytesPipe } from './convert-bytes';
+import { ConvertBytesPipe } from './convert-bytes/convert-bytes.pipe';
 
 @NgModule({
     declarations: [ConvertBytesPipe],
