+++ libs/platform/src/lib/utils/pipes/public_api.ts
@@ -1,0 +1,2 @@
+export * from './pipe.module';
+export * from './convert-bytes/convert-bytes.pipe';
